Move env_valid to ef_conf_utils
