+++ tests/unit_tests/test_ef_conf_utils.py
@@ -97,3 +97,53 @@ class TestEFConfUtils(unittest.TestCase):
     with self.assertRaises(ValueError) as exception:
       ef_conf_utils.global_env_valid(None)
     self.assertTrue("Invalid global env" in exception.exception.message)
+
+  def test_env_valid(self):
+    """
+    Checks if env_valid returns true for correctly named environments
+
+    Returns:
+      None
+
+    Raises:
+      AssertionError if any of the assert checks fail
+    """
+    for env in EFConfig.ENV_ACCOUNT_MAP:
+      # Attach a numeric value to environments that are ephemeral
+      if env in EFConfig.EPHEMERAL_ENVS:
+         env += '0'
+      self.assertTrue(ef_conf_utils.env_valid(env))
+
+    # Do tests for global and mgmt envs, which have a special mapping, Example: global.account_alias
+    if "global" in EFConfig.ENV_ACCOUNT_MAP:
+      for account_alias in EFConfig.ENV_ACCOUNT_MAP.values():
+        self.assertTrue(ef_conf_utils.env_valid("global." + account_alias))
+    if "mgmt" in EFConfig.ENV_ACCOUNT_MAP:
+      for account_alias in EFConfig.ENV_ACCOUNT_MAP.values():
+        self.assertTrue(ef_conf_utils.env_valid("mgmt." + account_alias))
+
+  def test_env_valid_invalid_envs(self):
+    """
+    Checks if env_valid returns ValueError for incorrectly name environments
+
+    Returns:
+      None
+
+    Raises:
+      AssertionError if any of the assert checks fail
+    """
+    # Create junk environment values by attaching numbers to non-ephemeral environments and not attaching numbers
+    # to ephemeral environments
+    for env in EFConfig.ENV_ACCOUNT_MAP:
+      if env not in EFConfig.EPHEMERAL_ENVS:
+        env += '0'
+      with self.assertRaises(ValueError):
+        ef_conf_utils.env_valid(env)
+
+    # Hard coded junk values
+    with self.assertRaises(ValueError):
+      ef_conf_utils.env_valid("invalid_env")
+    with self.assertRaises(ValueError):
+      ef_conf_utils.env_valid("")
+    with self.assertRaises(ValueError):
+      ef_conf_utils.env_valid(None)
